requirements.txt
torch
transformers
pillow
streamlit
scikit-learn
accelerate
bitsandbytes
sentencepiece
protobuf
pandas
